правки ДЗ Модуляция
