+++ js/main.js
@@ -1,6 +1,4 @@
 /* eslint-disable no-console */
-// eslint-disable-next-line no-unused-vars
-import {randomUserPhotos} from './util';
+import {randomUserPhotos} from './data';
 
-// eslint-disable-next-line eol-last
 console.log(randomUserPhotos);
